Google's patched-on LCOV support for coveragepy 4.
This is gross, it was done in a way that minized diffs.  Reworking this
to be non-ugly code for proper support in coveragepy would be great.

I'm making the PR on a branch of v4.5.x as our existing changes do not
apply cleanly after some of the refactoring in the git master (v5) code.

We're still using patched 4.4.2 internally at the time of PR creation.